AI_Service/requirements.txt
fastapi==0.115.12
uvicorn==0.30.6
python-multipart==0.0.20

groq==1.1.0
google-genai==1.66.0
pinecone-client==6.0.0
gTTS==2.5.4
python-dotenv==1.2.2
